move input to admin directory
