+++ src/App.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { Route, Switch } from 'react-router-dom';
 import Category from './user/pages/category'
 import Search from './user/pages/search'
-import AdminInput from './user/pages/input'
+import AdminInput from './admin/pages/input'
 import Detail from './user/pages/detail'
 import ViewAll from './user/pages/viewAll'
 import Top from './user/pages/top'
